# API_Document/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "如意外卖接口文档",
    "version": "1.0",
    "title": "如意外卖"
  },
  "host": "localhost:9000",
  "basePath": "/",
...
  "paths": {
...
    "/dish": {
      "post": {
        "tags": [
          "菜品管理"
        ],
        "summary": "新增菜品",
        "operationId": "saveUsingPOST_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "dishDto",
            "description": "dishDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Dish对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "菜品管理"
        ],
        "summary": "更新菜品信息",
        "operationId": "updateUsingPUT_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "dishDto",
            "description": "dishDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Dish对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "菜品管理"
        ],
        "summary": "删除菜品",
        "operationId": "deleteUsingDELETE_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dish/list": {
      "get": {
        "tags": [
          "菜品管理"
        ],
        "summary": "分类id查询菜品信息",
        "operationId": "listUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "菜品分类id",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "code",
            "in": "query",
            "description": "商品码",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "description",
            "in": "query",
            "description": "描述信息",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "image",
            "in": "query",
            "description": "图片",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "是否删除",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "name",
            "in": "query",
            "description": "菜品名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "description": "菜品价格",
            "required": false,
            "type": "number"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "顺序",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "status",
            "in": "query",
            "description": "0 停售 1 起售",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«Dish对象»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/dish/page": {
      "get": {
        "tags": [
          "菜品管理"
        ],
        "summary": "分页查询菜品信息",
        "operationId": "pageUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/dish/status/{statusType}": {
      "post": {
        "tags": [
          "菜品管理"
        ],
        "summary": "停用或者启用菜品",
        "operationId": "saleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "statusType",
            "in": "path",
            "description": "statusType",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/dish/{id}": {
      "get": {
        "tags": [
          "菜品管理"
        ],
        "summary": "通过id查询菜品的详细信息",
        "operationId": "getUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Dish对象»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee": {
      "post": {
        "tags": [
          "员工管理"
        ],
        "summary": "新增员工",
        "operationId": "saveUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "employee",
            "description": "employee",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "员工管理"
        ],
        "summary": "修改员工信息",
        "operationId": "updateUsingPUT_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "employee",
            "description": "employee",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/login": {
      "post": {
        "tags": [
          "员工管理"
        ],
        "summary": "员工登录",
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "employee",
            "description": "employee",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Employee对象»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/logout": {
      "post": {
        "tags": [
          "员工管理"
        ],
        "summary": "员工退出",
        "operationId": "logoutUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/page": {
      "get": {
        "tags": [
          "员工管理"
        ],
        "summary": "分页查询员工信息",
        "operationId": "pageUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/{id}": {
      "get": {
        "tags": [
          "员工管理"
        ],
        "summary": "获取单个员工信息",
        "operationId": "getByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Employee对象»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/setmeal/dish/{id}": {
      "get": {
        "tags": [
          "套餐管理"
        ],
        "summary": "移动端查看套餐下的菜品信息",
        "operationId": "dishUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«Dish对象»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DishFlavor对象": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "dishId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "口味名称",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        },
        "value": {
          "type": "string",
          "description": "口味数据list",
          "refType": null
        }
      },
      "title": "DishFlavor对象",
      "description": "菜品口味关系表"
    },
    "Dish对象": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id",
          "refType": null
        },
        "categoryName": {
          "type": "string",
          "refType": null
        },
        "code": {
          "type": "string",
          "description": "商品码",
          "refType": null
        },
        "copies": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "description": {
          "type": "string",
          "description": "描述信息",
          "refType": null
        },
        "flavors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DishFlavor对象"
          },
          "refType": "DishFlavor对象"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "image": {
          "type": "string",
          "description": "图片",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "菜品名称",
          "refType": null
        },
        "price": {
          "type": "number",
          "description": "菜品价格",
          "refType": null
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "description": "顺序",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "0 停售 1 起售",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        }
      },
      "title": "Dish对象",
      "description": "菜品管理"
    },
    "Employee对象": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "idNumber": {
          "type": "string",
          "description": "身份证号",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "姓名",
          "refType": null
        },
        "password": {
          "type": "string",
          "description": "密码",
          "refType": null
        },
        "phone": {
          "type": "string",
          "description": "手机号",
          "refType": null
        },
        "sex": {
          "type": "string",
          "description": "性别",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 0:禁用，1:正常",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        },
        "username": {
          "type": "string",
          "description": "用户名",
          "refType": null
        }
      },
      "title": "Employee对象",
      "description": "员工信息"
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean",
          "refType": null
        },
        "column": {
          "type": "string",
          "refType": null
        }
      },
      "title": "OrderItem"
    },
...
    "Page": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string",
          "refType": null
        },
        "current": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "optimizeCountSql": {
          "type": "boolean",
          "refType": null
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          },
          "refType": "OrderItem"
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "records": {
          "type": "array",
          "items": {
            "type": "object"
          },
          "refType": "object"
        },
        "searchCount": {
          "type": "boolean",
          "refType": null
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Page"
    },
...
    "R«Dish对象»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Dish对象"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Dish对象»"
    },
    "R«Employee对象»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Employee对象"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Employee对象»"
    },
...
    "R«List«Dish对象»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dish对象"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«Dish对象»»"
    },
...
    "R«Page»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Page»"
    },
...
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
...
  }
}